bot/requirements.txt
python-telegram-bot[job-queue]
tinydb
py-mon
